(BSR)[PRO] refactor: move features reducer to TS
