+++ pro/src/hooks/useActiveFeature.ts
@@ -1,9 +1,12 @@
 import { useSelector } from 'react-redux'
 
 import { isFeatureActive } from 'store/features/selectors'
+import { RootState } from 'store/reducers'
 
 const useActiveFeature = (featureName: string): boolean => {
-  const isActive = useSelector(state => isFeatureActive(state, featureName))
+  const isActive = useSelector((state: RootState) =>
+    isFeatureActive(state, featureName)
+  )
 
   return isActive
 }
