--- pro/src/store/features/actions.js
@@ -1,11 +1,0 @@
-export const SET_FEATURES_IS_INITIALIZED = 'SET_FEATURE_IS_INITIALIZED'
-export const SET_FEATURES = 'SET_FEATURES'
-
-export const setIsInitialized = () => ({
-  type: SET_FEATURES_IS_INITIALIZED,
-})
-
-export const setFeatures = features => ({
-  features,
-  type: SET_FEATURES,
-})
